CSS/post: convert flexbox to grid for post content and ToC
Before using the grid, we used flexbox for laying out the Table of
Contents beside the content of the post on larger screen sizes.
However this resulted in some magic numbers and broken spacing at
different zoom levels.

Using a grid, albeit with an hack, ensures the content is properly
placed.

To place the ToC before the content at small screen sizes we dynamically
create two grid rows and place the items accordingly.